[mem][data] Generate DoubleSeriesStore from Int
- copy the generator code from `common/gen` with modification
to replace more than one string
- `IntSeriesStore` tests are passing, but there is no test for
DoubleSeriesStore yet.
- [ ] TODO: the SeriesStore need redsign, it should contains SeriesMeta
and many data blocks with some aggregated value instead of a large slice

+++ pkg/common/series.go
@@ -56,7 +56,7 @@ type RawSeries struct {
 	Points     json.RawMessage `json:"points"`
 }
 
-func SeriesTypeString(seriesType int) string {
+func SeriesTypeString(seriesType int64) string {
 	switch seriesType {
 	case TypeIntSeries:
 		return "int"
